scripts/requirements.txt
afp-sdk @ git+https://github.com/autonity/afp-sdk@v0.6.0-rc.8
